requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
networkx>=3.1
